requirements.txt
aiogram==2.25.2
aiohttp==3.9.3
Pillow==10.2.0
requests==2.31.0
exif==1.6.0
